Breaks out the BasicAuthProvider into a separate class
